Update README.md
Background: In this assignment you will be modeling the account balances of an investor, under a number of di erent investment scenarios. Some of these assume a xed rate of return continuous. In this case, the investment change from one year to the next is

Bi+1 = Bie

where is the decimal form of the rate. (So 5% is = 0:05.)

Other scenarios assume a rate of return that is lognormal with mean and standard deviation . In this case, the investment change from one year to the next is

Bi+1 = BieN( ; )

where N( ; ) is a random value from a normal distribution with mean and standard deviation .

    Suppose $50,000 is invested on the investor’s 40th birthday at a xed rate of return = 7:6% con-tinuous. Determine the account balance each year from until the investor’s 65th birthday, and print a table that shows the investor’s age and the investment balance each year, with the balance given to the nearest cent. The rst three table entries are shown below.

        50000.00

        53948.13

        58208.01